fix(cli): split `forceRefreshPrompt` by platform to improve compatibility
- Moved `forceRefreshPrompt` into platform-specific files: `signal_unix.go` and `signal_windows.go`.
- Added a no-op implementation for Windows, aligning with the absence of `SIGWINCH` signal support.
- Cleaned up unused imports and the previous cross-platform implementation in `cli.go`.
